Add test suite for Checked C.
This adds a new test suite for Checked C to LLVM.  It is cribbed from the
CMake file and LIT configuration files for the clang test suite.

The Checked C tests will live in the Checked C repo.  This is so that the
tests can be re-used easily across different compiler implementations
and so that we can use the MIT license for the tests.  To drop the tests
into LLVM, simply clone the CheckedC repo to projects/checkedc-llvm/CheckedC.

I created a simple wrapper directory checkedc-llvm that is meant
to hold just build and test configuration files.  Those files are subject
to the LLVM license, so we are not putting them into the Checked C repo.

The Checked C tests are added as the build target check-checkedc.
If the CheckedC directory is not present under checkedc-llvm, no target is
added.

For checkedc-llvm/lit.cfg, I deleted options for clang tests that
weren't applicable and caused the script to break.  I also deleted the
code for doing out-of-tree builds.  That's not an interesting scenario at
this point.  I renamed the configuration variable clang_obj_dir to
checkedc_obj_dir, to avoid a potentially confusing name down the road.
(the object files are being dropped under the checkedc-llvm directory in
the object directory, not under the clang/test directory).