Add clang and CheckedC as submodules for the master branch.
This change adds the checkedc-clang and checkedc repos as submodules of the
checkedc-llvm repo for the master branch.  While there is some debate on the
Internet about the usefulness of git submodules, submodules seem to be a
good way to sensible coordinate repos that must be synchronized for tests
and builds to work.

For the checkedc repo, the tests branch is used as the submodule of the
master branch for now.  This is because the master branch mirrors what is on
Github and tests haven't been made available there.