Add clang and CheckedC as submodules.
This change adds the checkedc-clang and checkedc repos as submodules of the
checkedc-llvm repo.  While there is some debate on the Internet about the
usefulness of git submodules, submodules seem to be a good way to sensibly
coordinate repos that have a parent-child relationship and that must be
synchronized for tests and builds to work.

For the checkedc repo, the tests branch is used as the submodule of the
master branch for now.   This is because the master branch for CheckedC
mirrors what is on Github and tests haven't been made available there.

Recall that branches are really just pointers to changesets (which are
represented using SHA1 hashes), so we don't need to concern ourselves with
specifying a branch for the submodule.  All Git will care about is having the
right changeset.