Rename CheckedC repo to checkedc.
Tests for Checked C live in the Checked C repo, which is placed
under projects/checkedc-wrapper.  The Checked C used to be named
CheckedC.  We've made the name of the repo lowercase to match
the names of the LLVM and clang repos for Checked C.  A few
places that refer to the old name need to be fixed up.