Make CHECKEDC_IN_TREE globally defined.
Early versions of CMake used to make variables globally scoped.  Now
CMake uses dynamically-scoped variables, where a new scope is created at
the directory level and the variables are copied into the scope from the
parent scope.  This meant that setting the variable in nested CMake file
had no global effect.

Move setting the variable to the root CMake file for LLVM, so that the value
is globally visible.

This patch was contributed by Sam Elliott.