Copy Checked C headers to clang system header directory. (#7)
The Checked C header files are stored in the Checked C repo.   We want to copy them to the clang system header directory.  The Checked C repo is stored under llvm\projects\checked-wrapper, not clang.  This change causes the cmake files for LLVM to pick up the cmake file for the Checked C header files.  That adds a new target that can be used by the clang build system.