Integration with VS 2017. (#9)
Add support for integrating clang as a platform toolset for Visual Studio 2017.

In prior versions of Visual Studio, MSBuild was installed separately in its own directory.  In VS 2017, it is installed under the VS 2017 directory.    Use the vswhere program (https://github.com/Microsoft/vswhere) to location the installation of VS 2017.    This program is guaranteed to be placed in a known location, and can then be used to locate VS 2017.

From there, we can install .props and .targets file for MSBuild.   These are slightly modified versions of the (misnamed) VS 2014 files.

This LLVM repo was set up to have submodules for checkedc and clang.   This change also includes updates for those submodules (even though we don't pay much attention to keeping hashes up-to-date for them, as they would involve very frequent commits).