VS2017 installer part2 (#10)
On 32-bit OSes before Windows 10, vswhere is placed under %ProgramFiles%, not %ProgramFiles(x86)%.   Add code to handle this situation.

Visual Studio allows side-by-side installations, including installations of different products (Community, Professional, and Enterprise) with the same version and the same product with the same version.

For now, just use the latest version of VS 2017 installed in the install and uninstall scripts, so we get predictable behavior.   Alternately, we could go ahead and update every single version of VS 2017 installed on the machine.  That seems a little aggressive, since if someone is using side-by-side versions of VS there may be a reason.