Update location of registry key for Checked C LLVM/clang install. (#12)
This addresses Microsoft/checkedc-clang#442 (Visual Studio integration broken installer changes). We install the Checked C version of LLVM in a different location than the main-line version, to avoid overwriting an existing installation. I updated the organization name, which caused the location of a registry key used the MS Build property files to change. Update the location of the registry key in those files.