Fix lit testing using VS 2017. (#13)
When clang is run under the lit testing harness on a Windows machine with
only VS 2017 installed, clang can't find the VS install directory.  This
can cause tests that do linking to fail.   The lit scripts are filtering
out the ProgramData environment variable.   Clang is using a VS 2017 library
for finding the VS 2017 install location, and the library depends on the
ProgramData environment variable being set correctly.  The fix is to not
filter out the ProgramData variable.