Fixed AssertOK() of ICmpInst for _MMSafe_ptr.
When a _MMSafe_ptr is one of the operands of an ICmpInst, its inner
pointer should be extracted for the comparison; otherwise the compiler
will end up using the whole struct reppresentation of the _MMSafe_ptr
to compare with the other operand, which leads to an assertion failure
because a struct type cannot be used in an ICmpInst.

This commit only solves the false alarm raised by AssertOK(); it does
not solve the whole ICmpInst with _MMSafe_ptr problem: currently the
compiler uses "Value*" type as the operands when creating an ICmpInst
(check the "CreateICmp()" method in include/llvm/IR/IRBuilder.h").
The Value contains the struct of a _MMSafe_ptr; so later in the compile
process we need extract the real pointer for comparison.