Generate LLVM IR code for _MMSafe_ptr. (#21)
* Added code of building a _MMSafe_ptr type in LLVM IR.

Added a function to allocate a struct for a _MMSafe_ptr pointer.
See the comment of commit 9f8666b of the checkedc-clang repo for
more details.

9f8666b: https://github.com/microsoft/checkedc-clang/commit/9f8666bae94b782d8ee49d058a50a6d850b252c5

* Added a type check mthod for _MMSafe_ptr on the LLVM IR side.

Added a field and a function for llvm::PointerType class to check
if a PointerType is a _MMSafe_ptr<T> type.

* Added a method of llvm::StructType to check if it is a _MMSafe_ptr.

Since we implement _MMSafe_ptr as a struct, it's useful to have
a method to check if a llvm::StructType represents a _MMSafe_ptr.

* Added methods to extract the inner pointer inside a _MMSafe_ptr.

Added one method for llvm::Type and one method for llvm::StructType
to extract the inner pointer encoded inside a _MMSafe_ptr if the
target type is a _MMSafe_ptr. This two methods are very useful
during LLVM IR code generation.

* Fixed AssertOK() of ICmpInst for _MMSafe_ptr.

When a _MMSafe_ptr is one of the operands of an ICmpInst, its inner
pointer should be extracted for the comparison; otherwise the compiler
will end up using the whole struct reppresentation of the _MMSafe_ptr
to compare with the other operand, which leads to an assertion failure
because a struct type cannot be used in an ICmpInst.

This commit only solves the false alarm raised by AssertOK(); it does
not solve the whole ICmpInst with _MMSafe_ptr problem: currently the
compiler uses "Value*" type as the operands when creating an ICmpInst
(check the "CreateICmp()" method in include/llvm/IR/IRBuilder.h").
The Value contains the struct of a _MMSafe_ptr; so later in the compile
process we need extract the real pointer for comparison.

* Changed the return type of methods of getting inner ptr of _MMSafe_ptr.

Changed the return type of the two methods of getting the inner pointer
of a _MMSafe_ptr from "Type *" to "PointerType *" because we know it
is actually a PointerType. This saves us the effort of casting "Type *"
to "PointerType *" when these two methods are used.

* Fixed an indentation problem.

* Fixed a typo.